feat: @solana/wasm refactor code
