+++ wasm/types/Cargo.toml.d.ts
@@ -1,3 +1,4 @@
+// TODO: incorporate types from wasm-pack
 // export type Exports;// = typeof import('./../target/wasm-pack/solana-wasm/index');
 declare const init: () => Promise<any>;
 export default init;
